fix build ishtar, bump versions, fix commands

+++ tools/rune-cli/cmd/UpdateWorkloadCommand.cs
@@ -2,6 +2,7 @@ namespace vein.cmd;
 
 using System.ComponentModel;
 using System.Threading.Tasks;
+using System.Xml.Linq;
 using Spectre.Console;
 using Spectre.Console.Cli;
 using static UninstallWorkloadCommand;
@@ -11,7 +12,7 @@ public class UpdateWorkloadCommandSettings : CommandSettings
 {
     [Description("A package name.")]
     [CommandArgument(0, "[PACKAGE]")]
-    public required string PackageName { get; set; }
+    public required RunePackageKey PackageName { get; set; }
 
     [Description("Package version")]
     [CommandOption("--manifest", IsHidden = true)]
@@ -24,13 +25,18 @@ public class UpdateWorkloadCommand(ShardRegistryQuery query, ShardStorage storag
 
     public override async Task<int> ExecuteAsync(ProgressContext context, UpdateWorkloadCommandSettings settings)
     {
-        var uninstallResult = await new UninstallWorkloadCommand().ExecuteAsync(context,
-            new UninstallWorkloadCommandSettings()
-            {
-                PackageName = settings.PackageName.Split("@").First()
-            });
-        if (uninstallResult is not 0 and PACKAGE_NOT_FOUND)
-            return uninstallResult;
+        var packageName = settings.PackageName.Name;
+        var packageVersion = settings.PackageName.Version;
+        if (!WorkloadDirectory.SubDirectory(packageName).Exists)
+        {
+            var uninstallResult = await new UninstallWorkloadCommand().ExecuteAsync(context,
+                new UninstallWorkloadCommandSettings()
+                {
+                    PackageName = settings.PackageName.Name
+                });
+            if (uninstallResult is not 0 )
+                return uninstallResult;
+        }
         var installResult = await new InstallWorkloadCommand(query, storage, db).ExecuteAsync(context,
             new InstallWorkloadCommandSettings()
             {
